Move scheduling-related config definitions into pkg
