+++ pkg/mock/mockconfig/mockconfig.go
@@ -22,7 +22,7 @@ import (
 // NewTestOptions creates default options for testing.
 func NewTestOptions() *config.PersistOptions {
 	// register default schedulers in case config check fail.
-	for _, d := range config.DefaultSchedulers {
+	for _, d := range sc.DefaultSchedulers {
 		sc.RegisterScheduler(d.Type)
 	}
 	c := config.NewConfig()
