+++ pkg/schedule/config/util_test.go
@@ -1,4 +1,4 @@
-// Copyright 2018 TiKV Project Authors.
+// Copyright 2023 TiKV Project Authors.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -17,28 +17,40 @@ package config
 import (
 	"testing"
 
+	"github.com/pingcap/kvproto/pkg/metapb"
 	"github.com/stretchr/testify/require"
 )
 
-func TestValidateURLWithScheme(t *testing.T) {
+func TestValidateLabels(t *testing.T) {
 	re := require.New(t)
 	tests := []struct {
-		addr   string
+		label  string
 		hasErr bool
 	}{
-		{"", true},
-		{"foo", true},
-		{"/foo", true},
-		{"http", true},
-		{"http://", true},
-		{"http://foo", false},
-		{"https://foo", false},
-		{"http://127.0.0.1", false},
-		{"http://127.0.0.1/", false},
-		{"https://foo.com/bar", false},
-		{"https://foo.com/bar/", false},
+		{"z1", false},
+		{"z-1", false},
+		{"h1;", true},
+		{"z_1", false},
+		{"z_1&", true},
+		{"cn", false},
+		{"Zo^ne", true},
+		{"z_", true},
+		{"hos&t-15", true},
+		{"_test1", true},
+		{"-test1", true},
+		{"127.0.0.1", false},
+		{"www.pingcap.com", false},
+		{"h_127.0.0.1", false},
+		{"a", false},
+		{"a/b", false},
+		{"ab/", true},
+		{"/ab", true},
+		{"$abc", false},
+		{"$", true},
+		{"a$b", true},
+		{"$$", true},
 	}
 	for _, test := range tests {
-		re.Equal(test.hasErr, ValidateURLWithScheme(test.addr) != nil)
+		re.Equal(test.hasErr, ValidateLabels([]*metapb.StoreLabel{{Key: test.label}}) != nil)
 	}
 }
