+++ pkg/schedule/filter/filters.go
@@ -31,7 +31,7 @@ import (
 )
 
 // SelectSourceStores selects stores that be selected as source store from the list.
-func SelectSourceStores(stores []*core.StoreInfo, filters []Filter, conf config.SharedConfig, collector *plan.Collector,
+func SelectSourceStores(stores []*core.StoreInfo, filters []Filter, conf config.SharedConfigProvider, collector *plan.Collector,
 	counter *Counter) []*core.StoreInfo {
 	return filterStoresBy(stores, func(s *core.StoreInfo) bool {
 		return slice.AllOf(filters, func(i int) bool {
@@ -55,7 +55,7 @@ func SelectSourceStores(stores []*core.StoreInfo, filters []Filter, conf config.
 }
 
 // SelectUnavailableTargetStores selects unavailable stores that can't be selected as target store from the list.
-func SelectUnavailableTargetStores(stores []*core.StoreInfo, filters []Filter, conf config.SharedConfig,
+func SelectUnavailableTargetStores(stores []*core.StoreInfo, filters []Filter, conf config.SharedConfigProvider,
 	collector *plan.Collector, counter *Counter) []*core.StoreInfo {
 	return filterStoresBy(stores, func(s *core.StoreInfo) bool {
 		targetID := strconv.FormatUint(s.GetID(), 10)
@@ -85,7 +85,7 @@ func SelectUnavailableTargetStores(stores []*core.StoreInfo, filters []Filter, c
 }
 
 // SelectTargetStores selects stores that be selected as target store from the list.
-func SelectTargetStores(stores []*core.StoreInfo, filters []Filter, conf config.SharedConfig, collector *plan.Collector,
+func SelectTargetStores(stores []*core.StoreInfo, filters []Filter, conf config.SharedConfigProvider, collector *plan.Collector,
 	counter *Counter) []*core.StoreInfo {
 	if len(filters) == 0 {
 		return stores
@@ -133,9 +133,9 @@ type Filter interface {
 	Scope() string
 	Type() filterType
 	// Source Return plan.Status to show whether be filtered as source
-	Source(conf config.SharedConfig, store *core.StoreInfo) *plan.Status
+	Source(conf config.SharedConfigProvider, store *core.StoreInfo) *plan.Status
 	// Target Return plan.Status to show whether be filtered as target
-	Target(conf config.SharedConfig, store *core.StoreInfo) *plan.Status
+	Target(conf config.SharedConfigProvider, store *core.StoreInfo) *plan.Status
 }
 
 // comparingFilter is an interface to filter target store by comparing source and target stores
@@ -146,7 +146,7 @@ type comparingFilter interface {
 }
 
 // Target checks if store can pass all Filters as target store.
-func Target(conf config.SharedConfig, store *core.StoreInfo, filters []Filter) bool {
+func Target(conf config.SharedConfigProvider, store *core.StoreInfo, filters []Filter) bool {
 	storeID := strconv.FormatUint(store.GetID(), 10)
 	for _, filter := range filters {
 		status := filter.Target(conf, store)
@@ -189,14 +189,14 @@ func (f *excludedFilter) Type() filterType {
 	return excluded
 }
 
-func (f *excludedFilter) Source(conf config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f *excludedFilter) Source(conf config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	if _, ok := f.sources[store.GetID()]; ok {
 		return statusStoreAlreadyHasPeer
 	}
 	return statusOK
 }
 
-func (f *excludedFilter) Target(conf config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f *excludedFilter) Target(conf config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	if _, ok := f.targets[store.GetID()]; ok {
 		return statusStoreAlreadyHasPeer
 	}
@@ -219,11 +219,11 @@ func (f *storageThresholdFilter) Type() filterType {
 	return storageThreshold
 }
 
-func (f *storageThresholdFilter) Source(conf config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f *storageThresholdFilter) Source(conf config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	return statusOK
 }
 
-func (f *storageThresholdFilter) Target(conf config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f *storageThresholdFilter) Target(conf config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	if !store.IsLowSpace(conf.GetLowSpaceRatio()) {
 		return statusOK
 	}
@@ -287,11 +287,11 @@ func (f *distinctScoreFilter) Type() filterType {
 	return distinctScore
 }
 
-func (f *distinctScoreFilter) Source(_ config.SharedConfig, _ *core.StoreInfo) *plan.Status {
+func (f *distinctScoreFilter) Source(_ config.SharedConfigProvider, _ *core.StoreInfo) *plan.Status {
 	return statusOK
 }
 
-func (f *distinctScoreFilter) Target(_ config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f *distinctScoreFilter) Target(_ config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	score := core.DistinctScore(f.labels, f.stores, store)
 	switch f.policy {
 	case locationSafeguard:
@@ -348,9 +348,9 @@ func (f *StoreStateFilter) Type() filterType {
 
 // conditionFunc defines condition to determine a store should be selected.
 // It should consider if the filter allows temporary states.
-type conditionFunc func(config.SharedConfig, *core.StoreInfo) *plan.Status
+type conditionFunc func(config.SharedConfigProvider, *core.StoreInfo) *plan.Status
 
-func (f *StoreStateFilter) isRemoved(_ config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f *StoreStateFilter) isRemoved(_ config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	if store.IsRemoved() {
 		f.Reason = storeStateTombstone
 		return statusStoreRemoved
@@ -359,7 +359,7 @@ func (f *StoreStateFilter) isRemoved(_ config.SharedConfig, store *core.StoreInf
 	return statusOK
 }
 
-func (f *StoreStateFilter) isDown(conf config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f *StoreStateFilter) isDown(conf config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	if store.DownTime() > conf.GetMaxStoreDownTime() {
 		f.Reason = storeStateDown
 		return statusStoreDown
@@ -369,7 +369,7 @@ func (f *StoreStateFilter) isDown(conf config.SharedConfig, store *core.StoreInf
 	return statusOK
 }
 
-func (f *StoreStateFilter) isRemoving(_ config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f *StoreStateFilter) isRemoving(_ config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	if store.IsRemoving() {
 		f.Reason = storeStateOffline
 		return statusStoresRemoving
@@ -378,7 +378,7 @@ func (f *StoreStateFilter) isRemoving(_ config.SharedConfig, store *core.StoreIn
 	return statusOK
 }
 
-func (f *StoreStateFilter) pauseLeaderTransfer(_ config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f *StoreStateFilter) pauseLeaderTransfer(_ config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	if !store.AllowLeaderTransfer() {
 		f.Reason = storeStatePauseLeader
 		return statusStoreRejectLeader
@@ -387,7 +387,7 @@ func (f *StoreStateFilter) pauseLeaderTransfer(_ config.SharedConfig, store *cor
 	return statusOK
 }
 
-func (f *StoreStateFilter) slowStoreEvicted(conf config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f *StoreStateFilter) slowStoreEvicted(conf config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	if store.EvictedAsSlowStore() {
 		f.Reason = storeStateSlow
 		return statusStoreRejectLeader
@@ -396,7 +396,7 @@ func (f *StoreStateFilter) slowStoreEvicted(conf config.SharedConfig, store *cor
 	return statusOK
 }
 
-func (f *StoreStateFilter) slowTrendEvicted(_ config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f *StoreStateFilter) slowTrendEvicted(_ config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	if store.IsEvictedAsSlowTrend() {
 		f.Reason = storeStateSlowTrend
 		return statusStoreRejectLeader
@@ -405,7 +405,7 @@ func (f *StoreStateFilter) slowTrendEvicted(_ config.SharedConfig, store *core.S
 	return statusOK
 }
 
-func (f *StoreStateFilter) isDisconnected(_ config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f *StoreStateFilter) isDisconnected(_ config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	if !f.AllowTemporaryStates && store.IsDisconnected() {
 		f.Reason = storeStateDisconnected
 		return statusStoreDisconnected
@@ -414,7 +414,7 @@ func (f *StoreStateFilter) isDisconnected(_ config.SharedConfig, store *core.Sto
 	return statusOK
 }
 
-func (f *StoreStateFilter) isBusy(_ config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f *StoreStateFilter) isBusy(_ config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	if !f.AllowTemporaryStates && store.IsBusy() {
 		f.Reason = storeStateBusy
 		return statusStoreBusy
@@ -423,7 +423,7 @@ func (f *StoreStateFilter) isBusy(_ config.SharedConfig, store *core.StoreInfo) 
 	return statusOK
 }
 
-func (f *StoreStateFilter) exceedRemoveLimit(_ config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f *StoreStateFilter) exceedRemoveLimit(_ config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	if !f.AllowTemporaryStates && !store.IsAvailable(storelimit.RemovePeer, f.OperatorLevel) {
 		f.Reason = storeStateExceedRemoveLimit
 		return statusStoreRemoveLimit
@@ -432,7 +432,7 @@ func (f *StoreStateFilter) exceedRemoveLimit(_ config.SharedConfig, store *core.
 	return statusOK
 }
 
-func (f *StoreStateFilter) exceedAddLimit(_ config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f *StoreStateFilter) exceedAddLimit(_ config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	if !f.AllowTemporaryStates && !store.IsAvailable(storelimit.AddPeer, f.OperatorLevel) {
 		f.Reason = storeStateExceedAddLimit
 		return statusStoreAddLimit
@@ -441,7 +441,7 @@ func (f *StoreStateFilter) exceedAddLimit(_ config.SharedConfig, store *core.Sto
 	return statusOK
 }
 
-func (f *StoreStateFilter) tooManySnapshots(conf config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f *StoreStateFilter) tooManySnapshots(conf config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	if !f.AllowTemporaryStates && (uint64(store.GetSendingSnapCount()) > conf.GetMaxSnapshotCount() ||
 		uint64(store.GetReceivingSnapCount()) > conf.GetMaxSnapshotCount()) {
 		f.Reason = storeStateTooManySnapshot
@@ -451,7 +451,7 @@ func (f *StoreStateFilter) tooManySnapshots(conf config.SharedConfig, store *cor
 	return statusOK
 }
 
-func (f *StoreStateFilter) tooManyPendingPeers(conf config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f *StoreStateFilter) tooManyPendingPeers(conf config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	if !f.AllowTemporaryStates &&
 		conf.GetMaxPendingPeerCount() > 0 &&
 		store.GetPendingPeerCount() > int(conf.GetMaxPendingPeerCount()) {
@@ -462,7 +462,7 @@ func (f *StoreStateFilter) tooManyPendingPeers(conf config.SharedConfig, store *
 	return statusOK
 }
 
-func (f *StoreStateFilter) hasRejectLeaderProperty(conf config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f *StoreStateFilter) hasRejectLeaderProperty(conf config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	if conf.CheckLabelProperty(config.RejectLeader, store.GetLabels()) {
 		f.Reason = storeStateRejectLeader
 		return statusStoreRejectLeader
@@ -495,7 +495,7 @@ const (
 	fastFailoverTarget
 )
 
-func (f *StoreStateFilter) anyConditionMatch(typ int, conf config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f *StoreStateFilter) anyConditionMatch(typ int, conf config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	var funcs []conditionFunc
 	switch typ {
 	case leaderSource:
@@ -527,7 +527,7 @@ func (f *StoreStateFilter) anyConditionMatch(typ int, conf config.SharedConfig, 
 
 // Source returns true when the store can be selected as the schedule
 // source.
-func (f *StoreStateFilter) Source(conf config.SharedConfig, store *core.StoreInfo) (status *plan.Status) {
+func (f *StoreStateFilter) Source(conf config.SharedConfigProvider, store *core.StoreInfo) (status *plan.Status) {
 	if f.TransferLeader {
 		if status = f.anyConditionMatch(leaderSource, conf, store); !status.IsOK() {
 			return
@@ -544,7 +544,7 @@ func (f *StoreStateFilter) Source(conf config.SharedConfig, store *core.StoreInf
 
 // Target returns true when the store can be selected as the schedule
 // target.
-func (f *StoreStateFilter) Target(conf config.SharedConfig, store *core.StoreInfo) (status *plan.Status) {
+func (f *StoreStateFilter) Target(conf config.SharedConfigProvider, store *core.StoreInfo) (status *plan.Status) {
 	if f.TransferLeader {
 		if status = f.anyConditionMatch(leaderTarget, conf, store); !status.IsOK() {
 			return
@@ -588,7 +588,7 @@ func (f labelConstraintFilter) Type() filterType {
 }
 
 // Source filters stores when select them as schedule source.
-func (f labelConstraintFilter) Source(conf config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f labelConstraintFilter) Source(conf config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	if placement.MatchLabelConstraints(store, f.constraints) {
 		return statusOK
 	}
@@ -596,7 +596,7 @@ func (f labelConstraintFilter) Source(conf config.SharedConfig, store *core.Stor
 }
 
 // Target filters stores when select them as schedule target.
-func (f labelConstraintFilter) Target(_ config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f labelConstraintFilter) Target(_ config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	if placement.MatchLabelConstraints(store, f.constraints) {
 		return statusOK
 	}
@@ -638,7 +638,7 @@ func (f *ruleFitFilter) Type() filterType {
 	return ruleFit
 }
 
-func (f *ruleFitFilter) Source(_ config.SharedConfig, _ *core.StoreInfo) *plan.Status {
+func (f *ruleFitFilter) Source(_ config.SharedConfigProvider, _ *core.StoreInfo) *plan.Status {
 	return statusOK
 }
 
@@ -647,7 +647,7 @@ func (f *ruleFitFilter) Source(_ config.SharedConfig, _ *core.StoreInfo) *plan.S
 // the replaced store can match the source rule.
 // RegionA:[1,2,3], move peer1 --> peer2 will not allow, because it's count not match the rule.
 // but transfer role peer1 --> peer2, it will support.
-func (f *ruleFitFilter) Target(_ config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f *ruleFitFilter) Target(_ config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	if f.oldFit.Replace(f.srcStore, store) {
 		return statusOK
 	}
@@ -691,11 +691,11 @@ func (f *ruleLeaderFitFilter) Type() filterType {
 	return ruleLeader
 }
 
-func (f *ruleLeaderFitFilter) Source(_ config.SharedConfig, _ *core.StoreInfo) *plan.Status {
+func (f *ruleLeaderFitFilter) Source(_ config.SharedConfigProvider, _ *core.StoreInfo) *plan.Status {
 	return statusOK
 }
 
-func (f *ruleLeaderFitFilter) Target(_ config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f *ruleLeaderFitFilter) Target(_ config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	targetStoreID := store.GetID()
 	targetPeer := f.region.GetStorePeer(targetStoreID)
 	if targetPeer == nil && !f.allowMoveLeader {
@@ -747,11 +747,11 @@ func (f *ruleWitnessFitFilter) Type() filterType {
 	return ruleFit
 }
 
-func (f *ruleWitnessFitFilter) Source(_ config.SharedConfig, _ *core.StoreInfo) *plan.Status {
+func (f *ruleWitnessFitFilter) Source(_ config.SharedConfigProvider, _ *core.StoreInfo) *plan.Status {
 	return statusOK
 }
 
-func (f *ruleWitnessFitFilter) Target(_ config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f *ruleWitnessFitFilter) Target(_ config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	targetStoreID := store.GetID()
 	targetPeer := f.region.GetStorePeer(targetStoreID)
 	if targetPeer == nil {
@@ -769,7 +769,7 @@ func (f *ruleWitnessFitFilter) Target(_ config.SharedConfig, store *core.StoreIn
 
 // NewPlacementSafeguard creates a filter that ensures after replace a peer with new
 // peer, the placement restriction will not become worse.
-func NewPlacementSafeguard(scope string, conf config.SharedConfig, cluster *core.BasicCluster, ruleManager *placement.RuleManager,
+func NewPlacementSafeguard(scope string, conf config.SharedConfigProvider, cluster *core.BasicCluster, ruleManager *placement.RuleManager,
 	region *core.RegionInfo, sourceStore *core.StoreInfo, oldFit *placement.RegionFit) Filter {
 	if conf.IsPlacementRulesEnabled() {
 		return newRuleFitFilter(scope, cluster, ruleManager, region, oldFit, sourceStore.GetID())
@@ -780,7 +780,7 @@ func NewPlacementSafeguard(scope string, conf config.SharedConfig, cluster *core
 // NewPlacementLeaderSafeguard creates a filter that ensures after transfer a leader with
 // existed peer, the placement restriction will not become worse.
 // Note that it only worked when PlacementRules enabled otherwise it will always permit the sourceStore.
-func NewPlacementLeaderSafeguard(scope string, conf config.SharedConfig, cluster *core.BasicCluster, ruleManager *placement.RuleManager, region *core.RegionInfo, sourceStore *core.StoreInfo, allowMoveLeader bool) Filter {
+func NewPlacementLeaderSafeguard(scope string, conf config.SharedConfigProvider, cluster *core.BasicCluster, ruleManager *placement.RuleManager, region *core.RegionInfo, sourceStore *core.StoreInfo, allowMoveLeader bool) Filter {
 	if conf.IsPlacementRulesEnabled() {
 		return newRuleLeaderFitFilter(scope, cluster, ruleManager, region, sourceStore.GetID(), allowMoveLeader)
 	}
@@ -790,7 +790,7 @@ func NewPlacementLeaderSafeguard(scope string, conf config.SharedConfig, cluster
 // NewPlacementWitnessSafeguard creates a filter that ensures after transfer a witness with
 // existed peer, the placement restriction will not become worse.
 // Note that it only worked when PlacementRules enabled otherwise it will always permit the sourceStore.
-func NewPlacementWitnessSafeguard(scope string, conf config.SharedConfig, cluster *core.BasicCluster, ruleManager *placement.RuleManager,
+func NewPlacementWitnessSafeguard(scope string, conf config.SharedConfigProvider, cluster *core.BasicCluster, ruleManager *placement.RuleManager,
 	region *core.RegionInfo, sourceStore *core.StoreInfo, oldFit *placement.RegionFit) Filter {
 	if conf.IsPlacementRulesEnabled() {
 		return newRuleWitnessFitFilter(scope, cluster, ruleManager, region, oldFit, sourceStore.GetID())
@@ -819,14 +819,14 @@ func (f *engineFilter) Type() filterType {
 	return engine
 }
 
-func (f *engineFilter) Source(_ config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f *engineFilter) Source(_ config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	if f.constraint.MatchStore(store) {
 		return statusOK
 	}
 	return statusStoreNotMatchRule
 }
 
-func (f *engineFilter) Target(_ config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f *engineFilter) Target(_ config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	if f.constraint.MatchStore(store) {
 		return statusOK
 	}
@@ -862,14 +862,14 @@ func (f *specialUseFilter) Type() filterType {
 	return specialUse
 }
 
-func (f *specialUseFilter) Source(conf config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f *specialUseFilter) Source(conf config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	if store.IsLowSpace(conf.GetLowSpaceRatio()) || !f.constraint.MatchStore(store) {
 		return statusOK
 	}
 	return statusStoreNotMatchRule
 }
 
-func (f *specialUseFilter) Target(conf config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f *specialUseFilter) Target(conf config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	if !f.constraint.MatchStore(store) {
 		return statusOK
 	}
@@ -936,11 +936,11 @@ func (f *isolationFilter) Type() filterType {
 	return isolation
 }
 
-func (f *isolationFilter) Source(conf config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f *isolationFilter) Source(conf config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	return statusOK
 }
 
-func (f *isolationFilter) Target(_ config.SharedConfig, store *core.StoreInfo) *plan.Status {
+func (f *isolationFilter) Target(_ config.SharedConfigProvider, store *core.StoreInfo) *plan.Status {
 	// No isolation constraint to fit
 	if len(f.constraintSet) == 0 {
 		return statusStoreNotMatchIsolation
