+++ pkg/schedule/schedulers/balance_witness_test.go
@@ -36,7 +36,7 @@ type balanceWitnessSchedulerTestSuite struct {
 	tc     *mockcluster.Cluster
 	lb     Scheduler
 	oc     *operator.Controller
-	conf   config.SchedulerConfig
+	conf   config.SchedulerConfigProvider
 }
 
 func (suite *balanceWitnessSchedulerTestSuite) SetupTest() {
