+++ server/api/cluster_test.go
@@ -21,10 +21,11 @@ import (
 
 	"github.com/pingcap/kvproto/pkg/metapb"
 	"github.com/stretchr/testify/suite"
+
+	sc "github.com/tikv/pd/pkg/schedule/config"
 	tu "github.com/tikv/pd/pkg/utils/testutil"
 	"github.com/tikv/pd/server"
 	"github.com/tikv/pd/server/cluster"
-	"github.com/tikv/pd/server/config"
 )
 
 type clusterTestSuite struct {
@@ -70,7 +71,7 @@ func (suite *clusterTestSuite) TestCluster() {
 	suite.NoError(err)
 
 	c2 := &metapb.Cluster{}
-	r := config.ReplicationConfig{
+	r := sc.ReplicationConfig{
 		MaxReplicas:          6,
 		EnablePlacementRules: true,
 	}
@@ -98,7 +99,7 @@ func (suite *clusterTestSuite) testGetClusterStatus() {
 	suite.NoError(err)
 	suite.True(status.RaftBootstrapTime.After(now))
 	suite.False(status.IsInitialized)
-	suite.svr.SetReplicationConfig(config.ReplicationConfig{MaxReplicas: 1})
+	suite.svr.SetReplicationConfig(sc.ReplicationConfig{MaxReplicas: 1})
 	err = tu.ReadGetJSON(re, testDialClient, url, &status)
 	suite.NoError(err)
 	suite.True(status.RaftBootstrapTime.After(now))
